EPISODE 43 - Thread Subscriptions: Part 4

+++ forum/database/migrations/2020_01_28_225302_create_thread_subscriptions_table.php
@@ -16,7 +16,7 @@ class CreateThreadSubscriptionsTable extends Migration
         Schema::create('thread_subscriptions', function (Blueprint $table) {
             $table->increments('id');
             $table->unsignedInteger('user_id');
-            $table->unsignedInteger('thread_id');
+            $table->unsignedBigInteger('thread_id');
             $table->timestamps();
             $table->unique(['user_id', 'thread_id']);
